+ post like method

+++ src/Controller/PostController.ts
@@ -9,6 +9,7 @@ import { Types } from 'mongoose';
 
 import SaveDTO from '../DTO/Post/Save';
 import EditDTO from '../DTO/Post/Edit';
+import LikeDTO from '../DTO/Post/Like';
 
 import * as express from 'express';
 import * as multer from 'multer';
@@ -39,8 +40,10 @@ export default class UserController implements Routable
         });
 
         this.router.get('/getPosts/:userId', this.getPosts);
+        this.router.get('/getPost/:postId', this.getPost);
         this.router.post('/save', upload.single('media'), this.save);
         this.router.put('/edit/:postId', this.edit);
+        this.router.put('/like/:postId', this.like);
         this.router.delete('/delete/:postId', this.delete);
     }
 
@@ -102,7 +105,6 @@ export default class UserController implements Routable
             const newPost = new Post({
                 user: request.session.user.id,
                 media: newMedia._id,
-                title: data.title,
                 description: data.description
             });
 
@@ -110,7 +112,7 @@ export default class UserController implements Routable
                 throw(new Error('Failed to save Post'));
             }
 
-            //TODO: NOTIFY THE SUBSCRIBERS
+            //TODO: NOTIFY THE FOLLOWERS
 
             response.status(200).send({
                 _id: newPost._id,
@@ -118,7 +120,6 @@ export default class UserController implements Routable
                     url: newMedia.url,
                     mimetype: newMedia.mimetype
                 },
-                title: newPost.title,
                 description: newPost.description
             });
 
@@ -140,7 +141,7 @@ export default class UserController implements Routable
     {
         try {
 
-            const posts = await Post.aggregate([
+            let posts = await Post.aggregate([
                 {$match: {
                     user: new Types.ObjectId(request.params.userId)
                 }},
@@ -169,8 +170,73 @@ export default class UserController implements Routable
                 },
                 {$project: {
                     _id: 1,
-                    title: 1,
+                    media: {
+                        $ifNull: ['$media', null]
+                    },
+                    likes: 1
+                }}
+            ]);
+
+            if (request.session?.user?.id){
+                //@ts-ignore
+                posts = posts.map((post) =>{
+                    //@ts-ignore
+                    post.liked = post.likes.findIndex((like: any) => like.equals(request.session.user.id)) !== -1
+                    return (post);
+                });
+            }
+
+            response
+            .status(200)
+            .send({
+                posts: posts
+            });
+
+        } catch(error){
+
+            response
+            .status(error instanceof ServerException ? error.httpCode : 500)
+            .send({
+                errors: error instanceof ServerException ? error.messages : ['Internal server error']
+            })
+        }
+    }
+
+    getPost = async (request: express.Request, response: express.Response) =>
+    {
+        try {
+
+            let post = await Post.aggregate([
+                {$match: {
+                    _id: new Types.ObjectId(request.params.postId)
+                }},
+                {$lookup: {
+                    from: 'medias',
+                    let: {'mediaId': '$media'},
+                    pipeline: [
+                        {$match: {
+                            $expr: {
+                                $eq: ['$_id', '$$mediaId']
+                            }
+                        }},
+                        {$project: {
+                            _id : 0,
+                            mimetype: 1,
+                            url: 1
+                        }}
+                    ],
+                    as: 'media'
+                }},
+                {
+                    $unwind: {
+                        path : '$media',
+                        preserveNullAndEmptyArrays: true,
+                    }
+                },
+                {$project: {
+                    _id: 1,
                     description: 1,
+                    likes: 1,
                     media: {
                         $ifNull: ['$media', null]
                     },
@@ -178,10 +244,21 @@ export default class UserController implements Routable
                 }}
             ]);
 
+            if (!post?.length){
+                throw(new ServerException([`post ${request.params.postId} does not exist`], 400));
+            }
+
+            post = post[0];
+
+            if (request.session?.user?.id){
+                //@ts-ignore
+                post.liked = post.likes.findIndex((like: any) => like.equals(request.session.user.id)) !== -1
+            }
+
             response
             .status(200)
             .send({
-                posts: posts
+                post: post
             });
 
         } catch(error){
@@ -203,12 +280,18 @@ export default class UserController implements Routable
             }
 
             const data = request.body;
-            if (data.title === undefined && data.description === undefined){
-                throw(new ServerException(['you must include one of the following parameters: title, description'], 400));
+            if (data.description === undefined){
+                throw(new ServerException(['you must include one of the following parameters: description'], 400));
             }
 
-            if (!await Post.exists({_id: request.params.postId})){
+            const post = await Post.findById(request.params.postId);
+
+            if (!post){
                 throw(new ServerException([`post ${request.params.postId} does not exist`], 400));
+            }
+
+            if (!post.user.equals(request.session.user.id)){
+                throw(new ServerException(['Prohibited'], 403));
             }
 
             const errors = await this.validateDTO(EditDTO, data);
@@ -217,8 +300,53 @@ export default class UserController implements Routable
                 throw(new ServerException(errors, 400));
             }
 
-            if (!await Post.updateOne({_id: request.params.postId}, {$set: data})){
-                throw(new Error(`Failed to updateOne Post with _id ${request.params.postId}`));
+            if (!await Post.updateOne({_id: post._id}, {$set: data})){
+                throw(new Error(`Failed to updateOne Post with _id ${post._id}`));
+            }
+
+            response.status(204).send()
+
+        } catch(error){
+
+            response
+            .status(error instanceof ServerException ? error.httpCode : 500)
+            .send({
+                errors: error instanceof ServerException ? error.messages : ['Internal server error']
+            })
+        }
+    }
+
+    like = async (request: express.Request, response: express.Response) =>
+    {
+        try {
+
+            if (!request.session?.user?.id){
+                throw(new ServerException(['Unauthorized'], 401));
+            }
+
+            const post = await Post.findById(request.params.postId);
+
+            if (!post){
+                throw(new ServerException([`post ${request.params.postId} does not exist`], 400));
+            }
+
+            const data = request.body;
+            const errors = await this.validateDTO(LikeDTO, data);
+
+            if (errors?.length){
+                throw(new ServerException(errors, 400));
+            }
+
+            const like = post.likes.includes(new Types.ObjectId(request.session.user.id));
+
+            if (data.like && !like){
+                if (!await Post.updateOne({_id: post._id}, {$push: {likes: request.session.user.id}})){
+                    throw(new Error(`Failed to updateOne Post with _id ${post._id}`));
+                }
+            } else if (!data.like && like){
+                if (!await Post.updateOne({_id: post._id}, {$pull: {likes: request.session.user.id}})){
+                    throw(new Error(`Failed to updateOne Post with _id ${post._id}`));
+                }
             }
 
             response.status(204).send()
@@ -245,6 +373,10 @@ export default class UserController implements Routable
 
             if (!post){
                 throw(new ServerException([`post ${request.params.postId} does not exist`], 400));
+            }
+
+            if (!post.user.equals(request.session.user.id)){
+                throw(new ServerException(['Prohibited'], 403));
             }
 
             if (!await Media.deleteOne({_id: post.media})){
